refactoring + fixes
