+++ package.json
@@ -20,8 +20,10 @@
     "babelify": "^10.0.0",
     "browserify": "16.2.3",
     "cheerio": "^0.22.0",
+    "cli-progress": "^3.6.0",
     "copy": "^0.3.2",
     "dirsum": "^0.1.1",
+    "glob": "^7.1.6",
     "handlebars": "^4.4.5",
     "hoek": "^6.1.3",
     "html-minifier": "^4.0.0",
@@ -46,11 +48,13 @@
     "ts-node": "^8.1.0",
     "typescript": "^3.4.5"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "tslint": "^6.1.0"
+  },
   "scripts": {
     "setup": "node steps/setup.js",
     "get": "node_modules/ts-node/dist/bin.js steps/getCategory.ts",
-    "transform": "node steps/transform.js",
+    "transform": "node_modules/ts-node/dist/bin.js steps/transform.ts",
     "export": "node_modules/typescript/bin/tsc && node_modules/browserify/bin/cmd.js res/js/index.js  -t babelify --presets [es2015] -o res/js/dist.js && node_modules/ts-node/dist/bin.js steps/export.ts",
     "start": "npm run setup && npm run get && npm run transform && npm run export",
     "test": "echo \"Error: no test specified\" && exit 1",
